1.调整FreeRTOS栈空间为50K；2.调整内存整理函数；3.调整指令缓冲区为256字节; 4.调整部分数据长度类型；5.调整每个任务的使用空间

+++ NewSmartChannel/2LowLayer/USART/bsp_uart_fifo.c
@@ -359,9 +359,9 @@ uint8_t comGetChar(COM_PORT_E _ucPort, uint8_t *_pByte)
 	return UartGetChar(pUart, _pByte);
 }
 
-uint8_t comRecvBuff(COM_PORT_E _ucPort,uint8_t *buf, uint8_t len)
+uint16_t comRecvBuff(COM_PORT_E _ucPort,uint8_t *buf, uint16_t len)
 {
-    uint8_t i = 0;
+    uint16_t i = 0;
 
     UART_T *pUart;
 	pUart = ComToUart(_ucPort);
